Calculate updated hypothesis weights for the closure of a graph

The weights created by graph_generate_weights() work immediately for Bonferroni testing, but parametric and Simes testing require additional calculations. The adjust_weights_*() functions apply parametric or Simes weight increases to get updated weights for testing. They also subset the weights columns by the appropriate groups

Usage

adjust_weights_parametric(weighting_strategy, corr, alpha, groups)

adjust_weights_simes(weighting_strategy, p, groups)

c_value_function(x, hypotheses, corr, alpha)

solve_c_parametric(hypotheses, corr, alpha)

Arguments

weighting_strategy

For parametric, a compact representation of graph_generate_weights() output, where missing hypotheses get a missing value for weights, and h-vectors are dropped. For Simes, just the weights from graph_generate_weights() output

corr

A numeric matrix of correlations between hypotheses' test statistics

alpha

A numeric scalar specifying the global significance level for testing

groups

A list of numeric vectors specifying hypotheses to test together

p

A numeric vector of p-values

x

The root to solve for with stats::uniroot()

hypotheses

A numeric vector of hypothesis weights

Value

Outputs:

For adjust_weights_*(), a matrix with the same shape as weighting_strategy, where the weights have been adjusted according to the specified adjustment method

For c_value_function(), the \(c_{J_h}\) value for the given group, according to Formula 6 of Xi et al. (2017).

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

p <- 1:6 / 200

g <- bonferroni_holm(6)
gw_large <- graph_generate_weights(g)

gw_0 <- gw_large[, 7:12]
gw <- ifelse(gw_large[, 1:6], gw_0, NA)

graphicalMCP:::adjust_weights_parametric(gw, diag(6), .05, list(1:3))
#>           H1        H2        H3
#> 1  0.1680752 0.1680752 0.1680752
#> 2  0.2020340 0.2020340 0.2020340
#> 3  0.2020340 0.2020340 0.2020340
#> 4  0.2531918 0.2531918 0.2531918
#> 5  0.2020340 0.2020340 0.2020340
#> 6  0.2531918 0.2531918 0.2531918
#> 7  0.2531918 0.2531918 0.2531918
#> 8  0.3390485 0.3390485 0.3390485
#> 9  0.2010082 0.2010082        NA
#> 10 0.2515823 0.2515823        NA
#> 11 0.2515823 0.2515823        NA
#> 12 0.3361584 0.3361584        NA
#> 13 0.2515823 0.2515823        NA
#> 14 0.3361584 0.3361584        NA
#> 15 0.3361584 0.3361584        NA
#> 16 0.5064113 0.5064113        NA
#> 17 0.2010082        NA 0.2010082
#> 18 0.2515823        NA 0.2515823
#> 19 0.2515823        NA 0.2515823
#> 20 0.3361584        NA 0.3361584
#> 21 0.2515823        NA 0.2515823
#> 22 0.3361584        NA 0.3361584
#> 23 0.3361584        NA 0.3361584
#> 24 0.5064113        NA 0.5064113
#> 25 0.2500000        NA        NA
#> 26 0.3333333        NA        NA
#> 27 0.3333333        NA        NA
#> 28 0.5000000        NA        NA
#> 29 0.3333333        NA        NA
#> 30 0.5000000        NA        NA
#> 31 0.5000000        NA        NA
#> 32 1.0000000        NA        NA
#> 33        NA 0.2010082 0.2010082
#> 34        NA 0.2515823 0.2515823
#> 35        NA 0.2515823 0.2515823
#> 36        NA 0.3361584 0.3361584
#> 37        NA 0.2515823 0.2515823
#> 38        NA 0.3361584 0.3361584
#> 39        NA 0.3361584 0.3361584
#> 40        NA 0.5064113 0.5064113
#> 41        NA 0.2500000        NA
#> 42        NA 0.3333333        NA
#> 43        NA 0.3333333        NA
#> 44        NA 0.5000000        NA
#> 45        NA 0.3333333        NA
#> 46        NA 0.5000000        NA
#> 47        NA 0.5000000        NA
#> 48        NA 1.0000000        NA
#> 49        NA        NA 0.2500000
#> 50        NA        NA 0.3333333
#> 51        NA        NA 0.3333333
#> 52        NA        NA 0.5000000
#> 53        NA        NA 0.3333333
#> 54        NA        NA 0.5000000
#> 55        NA        NA 0.5000000
#> 56        NA        NA 1.0000000
#> 57        NA        NA        NA
#> 58        NA        NA        NA
#> 59        NA        NA        NA
#> 60        NA        NA        NA
#> 61        NA        NA        NA
#> 62        NA        NA        NA
#> 63        NA        NA        NA
graphicalMCP:::adjust_weights_simes(gw_0, p, list(4:6))
#>           H4        H5        H6
#> 1  0.1666667 0.3333333 0.5000000
#> 2  0.2000000 0.4000000 0.4000000
#> 3  0.2000000 0.2000000 0.4000000
#> 4  0.2500000 0.2500000 0.2500000
#> 5  0.0000000 0.2000000 0.4000000
#> 6  0.0000000 0.2500000 0.2500000
#> 7  0.0000000 0.0000000 0.2500000
#> 8  0.0000000 0.0000000 0.0000000
#> 9  0.2000000 0.4000000 0.6000000
#> 10 0.2500000 0.5000000 0.5000000
#> 11 0.2500000 0.2500000 0.5000000
#> 12 0.3333333 0.3333333 0.3333333
#> 13 0.0000000 0.2500000 0.5000000
#> 14 0.0000000 0.3333333 0.3333333
#> 15 0.0000000 0.0000000 0.3333333
#> 16 0.0000000 0.0000000 0.0000000
#> 17 0.2000000 0.4000000 0.6000000
#> 18 0.2500000 0.5000000 0.5000000
#> 19 0.2500000 0.2500000 0.5000000
#> 20 0.3333333 0.3333333 0.3333333
#> 21 0.0000000 0.2500000 0.5000000
#> 22 0.0000000 0.3333333 0.3333333
#> 23 0.0000000 0.0000000 0.3333333
#> 24 0.0000000 0.0000000 0.0000000
#> 25 0.2500000 0.5000000 0.7500000
#> 26 0.3333333 0.6666667 0.6666667
#> 27 0.3333333 0.3333333 0.6666667
#> 28 0.5000000 0.5000000 0.5000000
#> 29 0.0000000 0.3333333 0.6666667
#> 30 0.0000000 0.5000000 0.5000000
#> 31 0.0000000 0.0000000 0.5000000
#> 32 0.0000000 0.0000000 0.0000000
#> 33 0.2000000 0.4000000 0.6000000
#> 34 0.2500000 0.5000000 0.5000000
#> 35 0.2500000 0.2500000 0.5000000
#> 36 0.3333333 0.3333333 0.3333333
#> 37 0.0000000 0.2500000 0.5000000
#> 38 0.0000000 0.3333333 0.3333333
#> 39 0.0000000 0.0000000 0.3333333
#> 40 0.0000000 0.0000000 0.0000000
#> 41 0.2500000 0.5000000 0.7500000
#> 42 0.3333333 0.6666667 0.6666667
#> 43 0.3333333 0.3333333 0.6666667
#> 44 0.5000000 0.5000000 0.5000000
#> 45 0.0000000 0.3333333 0.6666667
#> 46 0.0000000 0.5000000 0.5000000
#> 47 0.0000000 0.0000000 0.5000000
#> 48 0.0000000 0.0000000 0.0000000
#> 49 0.2500000 0.5000000 0.7500000
#> 50 0.3333333 0.6666667 0.6666667
#> 51 0.3333333 0.3333333 0.6666667
#> 52 0.5000000 0.5000000 0.5000000
#> 53 0.0000000 0.3333333 0.6666667
#> 54 0.0000000 0.5000000 0.5000000
#> 55 0.0000000 0.0000000 0.5000000
#> 56 0.0000000 0.0000000 0.0000000
#> 57 0.3333333 0.6666667 1.0000000
#> 58 0.5000000 1.0000000 1.0000000
#> 59 0.5000000 0.5000000 1.0000000
#> 60 1.0000000 1.0000000 1.0000000
#> 61 0.0000000 0.5000000 1.0000000
#> 62 0.0000000 1.0000000 1.0000000
#> 63 0.0000000 0.0000000 1.0000000