Convert between graphicalMCP, gMCP, and igraph graph classes

Graphs are handled with a different object in graphicalMCP, gMCP, and igraph. These functions convert minimally between the different classes.

Usage

as_initial_graph(graph)

# S3 method for graphMCP
as_initial_graph(graph)

# S3 method for igraph
as_initial_graph(graph)

as_graphMCP(graph)

# S3 method for initial_graph
as_graphMCP(graph)

as_igraph(graph)

# S3 method for initial_graph
as_igraph(graph)

Arguments

graph

An initial_graph object from graphicalMCP, a graphMCP object from gMCP, or an igraph object from igraph, depending on conversion type

Value

For as_graphMCP(), a gMCP graph object, for as_igraph(), an igraph object, and for as_initial_graph(), a graphicalMCP graph object

Details

Note that igraph and gMCP can set various attributes for vertices, edges, or a graph itself. These conversion functions only handle attributes related to names and weights. Other attributes will be dropped when converting.

Examples

g1 <- random_graph(5)

if (requireNamespace("gMCP", quietly = TRUE)) {
  g2 <- as_graphMCP(g1)

  all.equal(g1, as_initial_graph(g2))
}
#> [1] TRUE

if (requireNamespace("igraph", quietly = TRUE)) {
  g3 <- as_igraph(g1)

  all.equal(g1, as_initial_graph(g3))
}
#> igraph method
#>  [1] "names for target but not for current"                             
#>  [2] "Attributes: < Length mismatch: comparison on first 1 components >"
#>  [3] "Attributes: < Component “class”: 1 string mismatch >"             
#>  [4] "Length mismatch: comparison on first 2 components"                
#>  [5] "Component 1: names for target but not for current"                
#>  [6] "Component 1: Numeric: lengths (5, 1) differ"                      
#>  [7] "Component 2: Modes: numeric, logical"                             
#>  [8] "Component 2: Lengths: 25, 1"                                      
#>  [9] "Component 2: Attributes: < Modes: list, NULL >"                   
#> [10] "Component 2: Attributes: < Lengths: 2, 0 >"                       
#> [11] "Component 2: Attributes: < names for target but not for current >"
#> [12] "Component 2: Attributes: < current is not list-like >"            
#> [13] "Component 2: target is matrix, current is logical"