Test hypotheses with the adjusted significance method

Test hypotheses with the adjusted significance method

Usage

test_values_bonferroni(p, hypotheses, alpha, intersection = NA)

test_values_parametric(p, hypotheses, alpha, intersection = NA, test_corr)

test_values_simes(p, hypotheses, alpha, intersection = NA)

Arguments

p

A numeric vector of p-values

hypotheses

A numeric vector of hypothesis hypotheses

alpha

A numeric scalar specifying the global significance level for testing

intersection

A numeric scalar used to track which intersection the test values are from

test_corr

A numeric matrix of correlations between hypotheses' test statistics

Value

A data frame with columns specifying the values used to calculate each hypothesis test

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

w <- c(H1 = .5, H2 = .5, H3 = 0, H4 = 0)

p <- c(.024, .01, .026, .027)

graphicalMCP:::test_values_bonferroni(p, w, .05)
#>    Intersection Hypothesis       Test     p <= c_value * Weight * Alpha
#> H1           NA         H1 bonferroni 0.024 <=              0.5 *  0.05
#> H2           NA         H2 bonferroni 0.010 <=              0.5 *  0.05
#> H3           NA         H3 bonferroni 0.026 <=              0.0 *  0.05
#> H4           NA         H4 bonferroni 0.027 <=              0.0 *  0.05
#>    Inequality_holds
#> H1             TRUE
#> H2             TRUE
#> H3            FALSE
#> H4            FALSE
graphicalMCP:::test_values_parametric(p, w, .05, test_corr = diag(4))
#>    Intersection Hypothesis       Test     p <=  c_value * Weight * Alpha
#> H1           NA         H1 parametric 0.024 <= 1.012823 *    0.5 *  0.05
#> H2           NA         H2 parametric 0.010 <= 1.012823 *    0.5 *  0.05
#> H3           NA         H3 parametric 0.026 <= 1.012823 *    0.0 *  0.05
#> H4           NA         H4 parametric 0.027 <= 1.012823 *    0.0 *  0.05
#>    Inequality_holds
#> H1             TRUE
#> H2             TRUE
#> H3            FALSE
#> H4            FALSE
graphicalMCP:::test_values_simes(p, w, .05)
#>   Intersection Hypothesis  Test     p <= c_value * Weight * Alpha
#> 1           NA         H1 simes 0.024 <=              1.0 *  0.05
#> 2           NA         H2 simes 0.010 <=              0.5 *  0.05
#> 3           NA         H3 simes 0.026 <=              1.0 *  0.05
#> 4           NA         H4 simes 0.027 <=              1.0 *  0.05
#>   Inequality_holds
#> 1             TRUE
#> 2             TRUE
#> 3             TRUE
#> 4             TRUE