Obtain hypothesis rejection probabilities

It's often difficult to tell how likely a given hypothesis is to be rejected. This is where power simulations are useful. Under a set of distribution parameters, many p-values are generated, and the graph is tested against each one. Any testing strategy can be used. Then probabilities are calculated for each hypothesis to be rejected, as well as some additional probabilities such as expected rejections and probability of rejecting any hypothesis

Usage

calculate_power(
  graph,
  alpha = 0.05,
  test_groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  test_corr = NULL,
  sim_n = 100,
  marginal_power = rep(0, length(graph$hypotheses)),
  sim_corr = diag(length(graph$hypotheses)),
  sim_success = NULL,
  sim_seed = NULL,
  force_closure = FALSE
)

Arguments

graph

An initial graph as returned by create_graph()

alpha

A numeric scalar specifying the global significance level for testing

test_groups

A list of numeric vectors specifying hypotheses to test together

test_types

A character vector of tests to apply to the given groups

test_corr

Optional if no test_types are parametric. A numeric matrix of correlations between hypotheses' test statistics

sim_n

An integer scalar specifying how many simulations to run

marginal_power

A numeric vector of mean values to use when simulating p-values. Exactly one mean per hypothesis is needed, and p-values will be sampled from the multivariate normal distribution. See Details for more

sim_corr

A numeric matrix of correlations between hypotheses used to sample from the multivariate normal distribution to generate p-values

sim_success

A list of user-defined functions to apply to the power results. Functions must take one simulation's logical vector results as an input, and return a length-one logical vector. For instance, if "success" means rejecting hypotheses 1 and 2, you would use sim_success = list("1 and 2" = function(x) x[1] && x[2]). If the list is not named, the function body will be used as the name. Lambda functions also work, e.g. sim_success = list(\(x) x[3] || x[4])

sim_seed

(Optional) Random seed to set before simulating p-values. Set this to use a consistent set of p simulations across power calculations

force_closure

A logical scalar used to determine whether the full closure test should be used for Bonferroni testing. Ignored if any tests are non-Bonferroni

Value

A list with five elements

power_local - rejection proportion for each hypothesis individually

power_expected - average number of hypotheses rejected in a single simulation

power_at_least_1 - proportion of simulations which reject any hypothesis

power_all - proportion of simulations which reject all hypotheses

power_success - proportion of simulations which reject any of the hypotheses specified in sim_success

Details

The parameters of the normal distribution are set with marginal_power (means) and sim_corr (correlation between test statistics). The mean of each hypothesis should be set as its marginal power $$d_i=P_{\xi_i}(p_i\leq\alpha)$$ where \(\xi_i\) is the non-centrality parameter. The correlation between test statistics is induced by the study design.

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

par_gate <- simple_successive_1()

# the default is to test all hypotheses with: Bonferroni testing at alpha
# level .05, 0 mean under the alternative, and 0 correlation between
# hypotheses under the alternative
# the default of 100 simulations will usually need to be increased
calculate_power(par_gate, sim_n = 1e5)
#> 
#> Test parameters ----------------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.500000
#>   H2: 0.500000
#>   H3: 0.000000
#>   H4: 0.000000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 1.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 1.000000
#>   H3 0.000000 1.000000 0.000000 0.000000
#>   H4 1.000000 0.000000 0.000000 0.000000
#> 
#>   Global alpha = 0.05
#> 
#>   Test types
#>   bonferroni: (H1-H2-H3-H4)
#> 
#> Simulation parameters ----------------------------------------------------------
#>   Testing 100,000 simulations with multivariate normal params:
#> 
#>                   H1 H2 H3 H4
#>   Marginal power:  0  0  0  0
#> 
#>   Correlation:    H1 H2 H3 H4
#>                H1  1  0  0  0
#>                H2  0  1  0  0
#>                H3  0  0  1  0
#>                H4  0  0  0  1
#> 
#> Power calculation --------------------------------------------------------------
#>                                   H1      H2      H3    H4
#>        Power to reject each: 0.02439 0.02496 0.00056 6e-04
#> 
#>         Expected rejections: 0.05051
#>   Power to reject 1 or more: 0.04858
#>         Power to reject all: 0
#> 
#> Simulation details -------------------------------------------------------------
#>   p_sim_H1 p_sim_H2 p_sim_H3 p_sim_H4 rej_H1 rej_H2 rej_H3 rej_H4
#>   0.355060 0.985301 0.898615 0.746449  FALSE  FALSE  FALSE  FALSE
#>   0.164270 0.586517 0.505819 0.230204  FALSE  FALSE  FALSE  FALSE
#>   0.258887 0.749053 0.984107 0.130700  FALSE  FALSE  FALSE  FALSE
#>   0.204238 0.699309 0.171881 0.789199  FALSE  FALSE  FALSE  FALSE
#>   0.303678 0.773963 0.578348 0.118210  FALSE  FALSE  FALSE  FALSE
#>   0.081745 0.015699 0.328450 0.713742  FALSE   TRUE  FALSE  FALSE
#>   ...
#> 

# but any test group/type combination that works for [test_graph_closure()]
# can be used
calculate_power(
  par_gate,
  alpha = .025,
  test_groups = list(1:2, 3:4),
  test_types = c("s", "p"),
  test_corr = diag(4),
  sim_n = 1e5,
  sim_success = list(
    function(.) .[1] || .[2],
    function(.) .[1] && .[2]
  )
)
#> 
#> Test parameters ----------------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.500000
#>   H2: 0.500000
#>   H3: 0.000000
#>   H4: 0.000000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 1.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 1.000000
#>   H3 0.000000 1.000000 0.000000 0.000000
#>   H4 1.000000 0.000000 0.000000 0.000000
#> 
#>   Global alpha = 0.025
#> 
#>   Parametric testing correlation:    H1 H2 H3 H4
#>                                   H1  1  0  0  0
#>                                   H2  0  1  0  0
#>                                   H3  0  0  1  0
#>                                   H4  0  0  0  1
#> 
#>   Test types
#>        simes: (H1-H2)
#>   parametric: (H3-H4)
#> 
#> Simulation parameters ----------------------------------------------------------
#>   Testing 100,000 simulations with multivariate normal params:
#> 
#>                   H1 H2 H3 H4
#>   Marginal power:  0  0  0  0
#> 
#>   Correlation:    H1 H2 H3 H4
#>                H1  1  0  0  0
#>                H2  0  1  0  0
#>                H3  0  0  1  0
#>                H4  0  0  0  1
#> 
#> Power calculation --------------------------------------------------------------
#>                                   H1      H2      H3    H4
#>        Power to reject each: 0.01264 0.01216 0.00015 2e-04
#> 
#>         Expected rejections: 0.02515
#>   Power to reject 1 or more: 0.02455
#>         Power to reject all: 0
#> 
#>                                .[1] || .[2] .[1] && .[2]
#>       Probability of success:       0.02455      0.00025
#> 
#> Simulation details -------------------------------------------------------------
#>   p_sim_H1 p_sim_H2 p_sim_H3 p_sim_H4 rej_H1 rej_H2 rej_H3 rej_H4
#>   0.170638 0.749592 0.036396 0.432719  FALSE  FALSE  FALSE  FALSE
#>   0.872438 0.555653 0.573393 0.855057  FALSE  FALSE  FALSE  FALSE
#>   0.297114 0.510197 0.045849 0.326143  FALSE  FALSE  FALSE  FALSE
#>   0.193122 0.291148 0.850419 0.529134  FALSE  FALSE  FALSE  FALSE
#>   0.960844 0.417840 0.426725 0.271911  FALSE  FALSE  FALSE  FALSE
#>   0.652054 0.409873 0.897269 0.677025  FALSE  FALSE  FALSE  FALSE
#>   ...
#>