initial_graph object

Creates a list that represents an initial graphical multiple comparison procedure.

create_graph(hypotheses, transitions, names = NULL)

Arguments

hypotheses

A numeric vector of hypothesis weights in an initial graphical multiple comparison procedure. Must be a vector of values between 0 & 1 (inclusive). The length should match the row and column lengths of transitions. The sum of hypothesis weights cannot exceed 1

transitions

A numeric matrix of transition weights between hypotheses in an initial graphical multiple comparison procedure. Must be a square matrix of values between 0 & 1 (inclusive). The row and column lengths should match the length of hypotheses. Each row (Transition weights leaving a hypothesis) can sum to no more than 1. The diagonal (Transition weights from a hypothesis to itself) must be all 0s

names

(Optional) A character vector of hypothesis names. If not provided, names from hypotheses and transitions will be used. If names are not specified, hypotheses will be named sequentially as H1, H2, ...

Value

An S3 object of class initial_graph

Examples

# A graphical multiple comparison procedure with two primary hypotheses (H1
# and H2) and two secondary hypotheses (H3 and H4)
# See Figure 1 in Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer,
# W., & Rohmeyer, K. (2011). Graphical approaches for multiple comparison
# procedures using weighted Bonferroni, Simes, or parametric tests.
# Biometrical Journal, 53(6), 894-913.
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000

# Explicit names override names in `hypotheses` (with a warning)
hypotheses <- c(h1 = 0.5, h2 = 0.5, h3 = 0, h4 = 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
#> Warning: hypothesis names specified - overriding names in
#>                     'hypotheses' and 'transitions'
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000

# Explicit names override names in `transitions` (with a warning)
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  h1 = c(0, 0, 1, 0),
  h2 = c(0, 0, 0, 1),
  h3 = c(0, 1, 0, 0),
  h4 = c(1, 0, 0, 0)
)
names <- c("H1", "H2", "H3", "H4")
g <- create_graph(hypotheses, transitions, names)
#> Warning: hypothesis names specified - overriding names in
#>                     'hypotheses' and 'transitions'
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000

# Use names in `hypotheses`
hypotheses <- c(H1 = 0.5, H2 = 0.5, H3 = 0, H4 = 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000

# Use names in `transitions`
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  H1 = c(0, 0, 1, 0),
  H2 = c(0, 0, 0, 1),
  H3 = c(0, 1, 0, 0),
  H4 = c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000

# When names are not specified, hypotheses are numbered sequentially as
# H1, H2, ...
hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)
g
#> An MCP graph
#> 
#> --- Hypothesis weights ---
#> H1: (0.5000)
#> H2: (0.5000)
#> H3: (0.0000)
#> H4: (0.0000)
#> 
#> --- Transition weights ---
#>        H1     H2     H3     H4
#> H1 0.0000 0.0000 1.0000 0.0000
#> H2 0.0000 0.0000 0.0000 1.0000
#> H3 0.0000 1.0000 0.0000 0.0000
#> H4 1.0000 0.0000 0.0000 0.0000