Generate weights for each intersections hypothesis in the full closure tree of an MCP graph

Generate weights for each intersections hypothesis in the full closure tree of an MCP graph

generate_weights(graph)

generate_weights_recursive(graph)

gw_original(graph)

Arguments

graph

An MCP graph as created by create_graph()

Value

A numeric matrix of all intersection hypothesis weights. Each row corresponds to a single intersection hypothesis. The first half of the columns indicate which hypotheses are included in the given intersection hypothesis, and the second half of columns are the weights

Performance

Much thought was given to the performance of this code, as the memory and time usage can grow quickly as graph size grows. On the systems used for testing, a size 10 graph had a median run time of 20-60 ms. Run time increases at a rate of O(2 ^ n), so e.g. a size 5 graph takes approximately twice as long to run as a size 4 graph

Examples

par_gate <- create_graph(
  hypotheses = c(.5, .5, 0, 0),
  transitions = rbind(
    c(0, 0, 1, 0),
    c(0, 0, 0, 1),
    c(0, 1, 0, 0),
    c(1, 0, 0, 0)
  )
)

generate_weights(par_gate)
#>    H1 H2 H3 H4  H1  H2  H3  H4
#> 1   1  1  1  1 0.5 0.5 0.0 0.0
#> 2   1  1  1  0 0.5 0.5 0.0 0.0
#> 3   1  1  0  1 0.5 0.5 0.0 0.0
#> 4   1  1  0  0 0.5 0.5 0.0 0.0
#> 5   1  0  1  1 0.5 0.0 0.0 0.5
#> 6   1  0  1  0 1.0 0.0 0.0 0.0
#> 7   1  0  0  1 0.5 0.0 0.0 0.5
#> 8   1  0  0  0 1.0 0.0 0.0 0.0
#> 9   0  1  1  1 0.0 0.5 0.5 0.0
#> 10  0  1  1  0 0.0 0.5 0.5 0.0
#> 11  0  1  0  1 0.0 1.0 0.0 0.0
#> 12  0  1  0  0 0.0 1.0 0.0 0.0
#> 13  0  0  1  1 0.0 0.0 0.5 0.5
#> 14  0  0  1  0 0.0 0.0 1.0 0.0
#> 15  0  0  0  1 0.0 0.0 0.0 1.0