Perform graphical multiple comparison procedures efficiently for power calculation

These functions performs similarly to graph_test_closure() or graph_test_shortcut() but are optimized for efficiently calculating power. For example, generating weights and calculating adjusted weights can be done only once. Vectorization has been applied where possible.

Usage

graph_test_closure_fast(p, alpha, adjusted_weights, matrix_intersections)

graph_test_shortcut_fast(p, alpha, adjusted_weights)

Arguments

p

A numeric vector of one-sided p-values (unadjusted, raw), whose values should be between 0 & 1. The length should match the number of hypotheses in graph.

alpha

A numeric value of the one-sided overall significance level, which should be between 0 & 1. The default is 0.025 for one-sided hypothesis testing. Note that only one-sided tests are supported.

adjusted_weights

The adjusted hypothesis weights, which are the second half of columns from graph_generate_weights() output, adjusted by the appropriate test types (Bonferroni, Simes, or parametric).

matrix_intersections

A matrix of hypothesis indicators in a weighting strategy, which are the first half the graph_generate_weights() output.

Value

A logical or integer vector indicating whether each hypothesis can be rejected or not.

See also

Examples

set.seed(1234)
alpha <- 0.025
p <- c(0.018, 0.01, 0.105, 0.006)
num_hyps <- length(p)
g <- bonferroni_holm(rep(1 / 4, 4))
weighting_strategy <- graph_generate_weights(g)

matrix_intersections <- weighting_strategy[, seq_len(num_hyps), drop = FALSE]
adjusted_weights <-
  weighting_strategy[, -seq_len(num_hyps), drop = FALSE]

graphicalMCP:::graph_test_closure_fast(
  p,
  alpha,
  adjusted_weights,
  matrix_intersections
)
#>    H1    H2    H3    H4 
#> FALSE FALSE FALSE  TRUE 
graphicalMCP:::graph_test_shortcut_fast(
  p,
  alpha,
  adjusted_weights
)
#>    H1    H2    H3    H4 
#> FALSE FALSE FALSE  TRUE