Report details of hypothesis rejections

The slower graph testing functions have design choices made that favor ease of interpreting results over speed. Results include hypothesis rejection decisions of course, but also the test values that led to the final result. The functions include options for reporting details using the adjusted p-value method or critical value method.

Usage

test_graph_closure(
  graph,
  p,
  alpha = 0.025,
  groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  corr = NULL,
  verbose = FALSE,
  critical = FALSE
)

test_graph_shortcut(graph, p, alpha = 0.025, verbose = FALSE, critical = FALSE)

Arguments

graph

An initial graph as returned by create_graph()

p

A numeric vector of p-values

alpha

A numeric scalar specifying the global significance level for testing

groups

A list of numeric vectors specifying hypotheses to test together

test_types

A character vector of tests to apply to the given groups

corr

(Optional) A numeric matrix of correlations between hypotheses' test statistics

verbose

A logical scalar specifying whether the results for each intersection hypothesis should be included

critical

A logical scalar specifying whether hypothesis-level detail should be included in the results, including calculating critical values for parametric tests

Value

A graph_report object, a list of 4 elements: inputs, outputs, verbose, and critical

Inputs - A list of the input parameters used to run the test

Outputs - A list of global test results

Verbose - A matrix with detailed adjusted p-value results (graph deletion sequence for shortcut testing)

Critical - A data frame with hypothesis-level test details for each intersection (each step for shortcut testing)

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

See also

Examples

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)

g <- create_graph(hypotheses, transitions)
p <- c(.01, .005, .015, .022)

corr <- matrix(nrow = 4, ncol = 4)
corr[3:4, 3:4] <- .9
diag(corr) <- 1

corr2 <- matrix(.5, nrow = 4, ncol = 4)
diag(corr2) <- 1

# The default is all Bonferroni with alpha = .05
test_graph_closure(g, p)
#> 
#> Test parameters ----------------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.500000
#>   H2: 0.500000
#>   H3: 0.000000
#>   H4: 0.000000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 1.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 1.000000
#>   H3 0.000000 1.000000 0.000000 0.000000
#>   H4 1.000000 0.000000 0.000000 0.000000
#> 
#>   Alpha = 0.025
#> 
#>                          H1    H2    H3    H4
#>   Unadjusted p-values: 0.01 0.005 0.015 0.022
#> 
#>   Test types
#>   bonferroni: (H1-H2-H3-H4)
#> 
#> Test summary -------------------------------------------------------------------
#>   Hypothesis Adj. P-value Reject
#>           H1         0.02   TRUE
#>           H2         0.01   TRUE
#>           H3         0.03  FALSE
#>           H4         0.03  FALSE
#> 
#>   Updated graph after rejections
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.000000
#>   H2: 0.000000
#>   H3: 0.500000
#>   H4: 0.500000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 0.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 0.000000
#>   H3 0.000000 0.000000 0.000000 1.000000
#>   H4 0.000000 0.000000 1.000000 0.000000

# But tests can be specified at the hypothesis-level
test_graph_closure(
  graph = g,
  p = p,
  alpha = .025,
  groups = list(1, 2, 3:4),
  test_types = c("bonferroni", "simes", "parametric"),
  corr = corr
)
#> 
#> Test parameters ----------------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.500000
#>   H2: 0.500000
#>   H3: 0.000000
#>   H4: 0.000000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 1.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 1.000000
#>   H3 0.000000 1.000000 0.000000 0.000000
#>   H4 1.000000 0.000000 0.000000 0.000000
#> 
#>   Alpha = 0.025
#> 
#>                          H1    H2    H3    H4
#>   Unadjusted p-values: 0.01 0.005 0.015 0.022
#> 
#>   Correlation matrix:    H1 H2  H3  H4
#>                       H1  1 NA  NA  NA
#>                       H2 NA  1  NA  NA
#>                       H3 NA NA 1.0 0.9
#>                       H4 NA NA 0.9 1.0
#> 
#>   Test types
#>   bonferroni: (H1)
#>        simes: (H2)
#>   parametric: (H3-H4)
#> 
#> Test summary -------------------------------------------------------------------
#>   Hypothesis Adj. P-value Reject
#>           H1     0.020000   TRUE
#>           H2     0.010000   TRUE
#>           H3     0.021546   TRUE
#>           H4     0.022000   TRUE
#> 
#>   Updated graph after rejections
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.000000
#>   H2: 0.000000
#>   H3: 0.000000
#>   H4: 0.000000
#> 
#>   --- Transition weights ---
#>            H1       H2       H3       H4
#>   H1 0.000000 0.000000 0.000000 0.000000
#>   H2 0.000000 0.000000 0.000000 0.000000
#>   H3 0.000000 0.000000 0.000000 0.000000
#>   H4 0.000000 0.000000 0.000000 0.000000