Delete hypotheses from a graph

It is not always obvious what a graph's weights will look like after deleting one or more hypotheses. While generate_weights() calculates all sub-graphs' hypothesis weights, update_graph() gives a more detailed view for a single set of deletions, including transition weights as well.

Usage

update_graph(graph, keep)

Arguments

graph

An initial graph as returned by create_graph()

keep

A logical or integer vector, denoting which hypotheses to keep/delete. An entry of FALSE (or 0) corresponds to a deletion, and TRUE (or 1) corresponds to keeping a hypothesis

Value

An object of class updated_graph with 3 elements

The initial graph object

The Boolean vector indicating which hypotheses are kept/deleted

The updated graph object with specific hypotheses deleted

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- create_graph(hypotheses, transitions)

# Delete the second hypothesis
update_graph(g, c(TRUE, FALSE, TRUE, TRUE))
#> Initial graph
#> 
#> --- Hypothesis weights ---
#> H1: 0.5000
#> H2: 0.5000
#> H3: 0.0000
#> H4: 0.0000
#> 
#> --- Transition weights ---
#>         H1     H2     H3     H4
#>  H1 0.0000 0.0000 1.0000 0.0000
#>  H2 0.0000 0.0000 0.0000 1.0000
#>  H3 0.0000 1.0000 0.0000 0.0000
#>  H4 1.0000 0.0000 0.0000 0.0000
#> 
#> --------------------------------------------------------------------------------
#> 
#> --- Hypotheses kept ---
#>    H1    H2   H3   H4
#>  TRUE FALSE TRUE TRUE
#> 
#> --------------------------------------------------------------------------------
#> 
#> Updated graph
#> 
#> --- Hypothesis weights ---
#> H1: 0.5000
#> H2: 0.0000
#> H3: 0.0000
#> H4: 0.5000
#> 
#> --- Transition weights ---
#>         H1     H2     H3     H4
#>  H1 0.0000 0.0000 1.0000 0.0000
#>  H2 0.0000 0.0000 0.0000 0.0000
#>  H3 0.0000 0.0000 0.0000 1.0000
#>  H4 1.0000 0.0000 0.0000 0.0000
# Equivalent
# update_graph(g, c(1, 0, 1, 1))