Obtain hypothesis rejection probabilities

It's often difficult to tell how likely a given hypothesis is to be rejected. This is where power simulations are useful. Under a set of distribution parameters, many p-values are generated, and the graph is tested against each one. Any testing strategy can be used. Then probabilities are calculated for each hypothesis to be rejected, as well as some additional probabilities such as expected rejections and probability of rejecting any hypothesis

Usage

graph_calculate_power(
  graph,
  alpha = 0.025,
  power_marginal = rep(alpha, length(graph$hypotheses)),
  test_groups = list(seq_along(graph$hypotheses)),
  test_types = c("bonferroni"),
  test_corr = rep(list(NA), length(test_types)),
  sim_n = 100,
  sim_corr = diag(length(graph$hypotheses)),
  sim_success = NULL,
  verbose = FALSE
)

Arguments

graph

An initial graph as returned by graph_create()

alpha

A numeric scalar specifying the global significance level for testing

power_marginal

A numeric vector of mean values to use when simulating p-values. Exactly one mean per hypothesis is needed, and p-values will be sampled from the multivariate normal distribution. See Details for more

test_groups

A list of numeric vectors specifying hypotheses to test together

test_types

A character vector of tests to apply to the given groups

test_corr

Optional if no test_types are parametric. A numeric matrix of correlations between hypotheses' test statistics

sim_n

An integer scalar specifying how many simulations to run

sim_corr

A numeric matrix of correlations between hypotheses used to sample from the multivariate normal distribution to generate p-values

sim_success

A list of user-defined functions to apply to the power results. Functions must take one simulation's logical vector of results as an input, and return a length-one logical vector. For instance, if "success" means rejecting hypotheses 1 and 2, use sim_success = list("1 and 2" = function(x) x[1] && x[2]). If the list is not named, the function body will be used as the name. Lambda functions also work starting with R 4.1, e.g. sim_success = list(\(x) x[3] || x[4])

verbose

A logical scalar specifying whether the full matrix of simulations and test results should be included in the output or not

Value

A list with three elements

inputs - A list of input parameters

power - A list of measures of how often hypotheses are rejected

power_local - Rejection proportion for each hypothesis individually

power_expected - Average number of hypotheses rejected in a single simulation

power_at_least_1 - Proportion of simulations which reject any hypothesis

power_all - Proportion of simulations which reject all hypotheses

power_success - Proportion of simulations which reject any of the hypotheses specified in sim_success

details - An optional list of datasets showing simulated p-values and results for each simulation

Details

The parameters of the normal distribution are set with power_marginal (means) and sim_corr (correlation between test statistics). The mean of each hypothesis should be set as its marginal power $$d_i=P_{\xi_i}(p_i\leq\alpha)$$ where \(\xi_i\) is the non-centrality parameter. The correlation between test statistics is induced by the study design.

Success

Success will mean something different for each trial, so there's a lot of flexibility in the sim_success parameter. However, this flexibility means there's very little validation of inputs. It's up to the user to make sure the function(s) passed mean what they think. From an implementation perspective, each function will be applied row-wise to the matrix of test results for the simulation, resulting in a sim_n length vector. The mean of this vector is returned as "Probability of success"

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

par_gate <- simple_successive_1()

# The default is to test all hypotheses with: Bonferroni testing at alpha
# level .025, 0 mean under the alternative, and 0 correlation between
# hypotheses under the alternative
# The default of 100 simulations will usually need to be increased
graph_calculate_power(par_gate, sim_n = 1e5)
#> 
#> Test parameters ($inputs) ------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2: 0.5
#>   H3: 0.0
#>   H4: 0.0
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1  0  0  1  0
#>   H2  0  0  0  1
#>   H3  0  1  0  0
#>   H4  1  0  0  0
#> 
#>   Alpha = 0.025
#> 
#>   Test types
#>   bonferroni: (H1, H2, H3, H4)
#> 
#> Simulation parameters ($inputs) ------------------------------------------------
#>   Testing 100,000 simulations with multivariate normal params:
#> 
#>                      H1    H2    H3    H4
#>   Marginal power: 0.025 0.025 0.025 0.025
#> 
#>   Correlation:    H1 H2 H3 H4
#>                H1  1  0  0  0
#>                H2  0  1  0  0
#>                H3  0  0  1  0
#>                H4  0  0  0  1
#> 
#> Power calculation ($power) -----------------------------------------------------
#>                                    H1      H2      H3      H4
#>                  Local power: 0.01201 0.01166 0.00016 0.00019
#> 
#>   Expected no. of rejections: 0.02402
#>    Power to reject 1 or more: 0.02349
#>          Power to reject all: 0
#> 

# But any test group/type combination that works for [graph_test_closure()]
# can be used
graph_calculate_power(
  par_gate,
  alpha = .025,
  test_groups = list(1:2, 3:4),
  test_types = c("s", "p"),
  test_corr = list(NA, diag(2)),
  sim_n = 1e5,
  sim_success = list(
    function(.) .[1] || .[2],
    function(.) .[1] && .[2]
  )
)
#> 
#> Test parameters ($inputs) ------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2: 0.5
#>   H3: 0.0
#>   H4: 0.0
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1  0  0  1  0
#>   H2  0  0  0  1
#>   H3  0  1  0  0
#>   H4  1  0  0  0
#> 
#>   Alpha = 0.025
#> 
#>   Parametric testing correlation:    H3 H4
#>                                   H3  1  0
#>                                   H4  0  1
#> 
#>   Test types
#>        simes: (H1, H2)
#>   parametric: (H3, H4)
#> 
#> Simulation parameters ($inputs) ------------------------------------------------
#>   Testing 100,000 simulations with multivariate normal params:
#> 
#>                      H1    H2    H3    H4
#>   Marginal power: 0.025 0.025 0.025 0.025
#> 
#>   Correlation:    H1 H2 H3 H4
#>                H1  1  0  0  0
#>                H2  0  1  0  0
#>                H3  0  0  1  0
#>                H4  0  0  0  1
#> 
#> Power calculation ($power) -----------------------------------------------------
#>                                    H1      H2      H3      H4
#>                  Local power: 0.01203 0.01215 0.00018 0.00021
#> 
#>   Expected no. of rejections: 0.02457
#>    Power to reject 1 or more: 0.02403
#>          Power to reject all: 0
#> 
#>   Success measure   Power
#>      .[1] || .[2] 0.02403
#>      .[1] && .[2] 0.00015
#>