Find alternate rejection orderings for shortcut testing

When using graph_test_shortcut(), there may be multiple hypotheses which can be rejected at each step. The specific hypothesis chosen is decided based on the minimum adjusted p-value. This function shows every other order that deletes the same hypotheses, where each hypothesis is still a valid deletion at the step it is chosen.

Usage

graph_rejection_orderings(shortcut_test_result)

Arguments

shortcut_test_result

A graph_report object as returned by graph_test_shortcut()

Value

A modified graph_report object containing all valid orderings for deleting the significant hypotheses

Examples

graph <- simple_successive_2()

short_res <- graph_test_shortcut(graph, c(.018, .01, .03, .004))

# Reject H1, H2, and H4
short_res$outputs$rejected
#>    H1    H2    H3    H4 
#>  TRUE  TRUE FALSE  TRUE 

# But these cannot be rejected in any order - For instance, H4 has 0 weight
# in the initial graph and cannot be rejected first
graph_rejection_orderings(short_res)
#> 
#> Test parameters ($inputs) ------------------------------------------------------
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2: 0.5
#>   H3: 0.0
#>   H4: 0.0
#> 
#>   --- Transition weights ---
#>       H1  H2  H3  H4
#>   H1 0.0 0.5 0.5 0.0
#>   H2 0.5 0.0 0.0 0.5
#>   H3 0.0 1.0 0.0 0.0
#>   H4 1.0 0.0 0.0 0.0
#> 
#>   Alpha = 0.025
#> 
#>                           H1    H2    H3    H4
#>   Unadjusted p-values: 0.018 0.010 0.030 0.004
#> 
#>   Test types
#>   bonferroni: (H1, H2, H3, H4)
#> 
#> Test summary ($outputs) --------------------------------------------------------
#>   Hypothesis Adj. P-value Reject
#>           H1         0.02   TRUE
#>           H2         0.02   TRUE
#>           H3         0.03  FALSE
#>           H4         0.02   TRUE
#> 
#>   Final updated graph after removing rejected hypotheses
#> 
#>   --- Hypothesis weights ---
#>   H1: NA
#>   H2: NA
#>   H3:  1
#>   H4: NA
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1 NA NA NA NA
#>   H2 NA NA NA NA
#>   H3 NA NA  0 NA
#>   H4 NA NA NA NA
#> 
#> Alternate rejection orderings ($valid_rejection_orderings) ---------------------
#> H2 H4 H1 
#>  2  4  1 
#> 
#> H2 H1 H4 
#>  2  1  4 
#> 

# Finally, intermediate graphs can be obtained by putting one of the
# orderings outputs into `graph_update()`
graph_update(graph, delete = c(2, 1, 4))
#> Initial graph
#> 
#> --- Hypothesis weights ---
#> H1: 0.5
#> H2: 0.5
#> H3: 0.0
#> H4: 0.0
#> 
#> --- Transition weights ---
#>      H1  H2  H3  H4
#>  H1 0.0 0.5 0.5 0.0
#>  H2 0.5 0.0 0.0 0.5
#>  H3 0.0 1.0 0.0 0.0
#>  H4 1.0 0.0 0.0 0.0
#> 
#> Updated graph after deleting hypotheses 2, 1, 4
#> 
#> --- Hypothesis weights ---
#> H1: NA
#> H2: NA
#> H3:  1
#> H4: NA
#> 
#> --- Transition weights ---
#>     H1 H2 H3 H4
#>  H1 NA NA NA NA
#>  H2 NA NA NA NA
#>  H3 NA NA  0 NA
#>  H4 NA NA NA NA
#> 
#> Deletion sequence ($intermediate_graphs) ---------------------------------------
#> 
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2: 0.5
#>   H3: 0.0
#>   H4: 0.0
#> 
#>   --- Transition weights ---
#>       H1  H2  H3  H4
#>   H1 0.0 0.5 0.5 0.0
#>   H2 0.5 0.0 0.0 0.5
#>   H3 0.0 1.0 0.0 0.0
#>   H4 1.0 0.0 0.0 0.0
#> 
#>     Step 1: Updated graph after removing hypothesis 2
#> 
#>     --- Hypothesis weights ---
#>     H1: 0.75
#>     H2:   NA
#>     H3: 0.00
#>     H4: 0.25
#> 
#>     --- Transition weights ---
#>              H1       H2       H3       H4
#>     H1 0.000000       NA 0.666667 0.333333
#>     H2       NA       NA       NA       NA
#>     H3 0.500000       NA 0.000000 0.500000
#>     H4 1.000000       NA 0.000000 0.000000
#> 
#>       Step 2: Updated graph after removing hypotheses 2, 1
#> 
#>       --- Hypothesis weights ---
#>       H1:  NA
#>       H2:  NA
#>       H3: 0.5
#>       H4: 0.5
#> 
#>       --- Transition weights ---
#>          H1 H2 H3 H4
#>       H1 NA NA NA NA
#>       H2 NA NA NA NA
#>       H3 NA NA  0  1
#>       H4 NA NA  1  0
#> 
#>         Step 3: Updated graph after removing hypotheses 2, 1, 4
#> 
#>         --- Hypothesis weights ---
#>         H1: NA
#>         H2: NA
#>         H3:  1
#>         H4: NA
#> 
#>         --- Transition weights ---
#>            H1 H2 H3 H4
#>         H1 NA NA NA NA
#>         H2 NA NA NA NA
#>         H3 NA NA  0 NA
#>         H4 NA NA NA NA
#> 
#>   Final updated graph after removing deleted hypotheses
#> 
#>   --- Hypothesis weights ---
#>   H1: NA
#>   H2: NA
#>   H3:  1
#>   H4: NA
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1 NA NA NA NA
#>   H2 NA NA NA NA
#>   H3 NA NA  0 NA
#>   H4 NA NA NA NA
#>