Delete hypotheses from a graph

It is not always obvious what a graph's weights will look like after deleting one or more hypotheses. While graph_generate_weights() calculates all sub-graphs' hypothesis weights, graph_update() gives a more detailed view for a single set of deletions, including transition weights.

Usage

graph_update(graph, delete)

Arguments

graph

An initial graph as returned by graph_create()

delete

A logical or integer vector, denoting which hypotheses to keep/delete. A logical vector must match the size of the graph, with one entry per hypothesis, and it results in "unordered mode," where selected hypotheses are deleted in sequential order. An integer vector can be any length, but must only contain valid hypothesis numbers (greater than 0, and less than or equal to the size of the graph). This will trigger "ordered mode," where selected hypotheses are deleted in the order they appear in delete

Value

An object of class updated_graph with 4 elements

The initial graph object

The updated graph object with specified hypotheses deleted

A numeric vector indicating which hypotheses were deleted

When using ordered mode, a list of intermediate graphs for each deletion step

References

Bretz, F., Maurer, W., Brannath, W., and Posch, M. (2009). A graphical approach to sequentially rejective multiple test procedures. Statistics in Medicine, 28(4), 586–604. https://doi.org/10.1002/sim.3495

Bretz, F., Maurer, W., and Hommel, G. (2011). Test and power considerations for multiple endpoint analyses using sequentially rejective graphical procedures. Statistics in Medicine, 30(13), 1489–1501. https://doi.org/10.1002/sim.3988

Bretz, F., Posch, M., Glimm, E., Klinglmueller, F., Maurer, W., and Rohmeyer, K. (2011). Graphical approaches for multiple comparison procedures using weighted Bonferroni, Simes, or parametric tests. Biometrical Journal, 53(6), 894–913. https://doi.org/10.1002/bimj.201000239

Lu, K. (2016). Graphical approaches using a Bonferroni mixture of weighted Simes tests. Statistics in Medicine, 35(22), 4041–4055. https://doi.org/10.1002/sim.6985

Xi, D., Glimm, E., Maurer, W., and Bretz, F. (2017). A unified framework for weighted parametric multiple test procedures. Biometrical Journal, 59(5), 918–931. https://doi.org/10.1002/bimj.201600233

Xi, D., and Bretz, F. (2019). Symmetric graphs for equally weighted tests, with application to the Hochberg procedure. Statistics in Medicine, 38(27), 5268–5282. https://doi.org/10.1002/sim.8375

Rohmeyer K, Klinglmueller F (2020). gMCP: Graph Based Multiple Test Procedures. R package version 0.8-15, https://cran.r-project.org/package=gMCP.

Examples

hypotheses <- c(0.5, 0.5, 0, 0)
transitions <- rbind(
  c(0, 0, 1, 0),
  c(0, 0, 0, 1),
  c(0, 1, 0, 0),
  c(1, 0, 0, 0)
)
g <- graph_create(hypotheses, transitions)

# Delete the second and third hypotheses in any order
graph_update(g, c(FALSE, TRUE, TRUE, FALSE))
#> Initial and final graphs -------------------------------------------------------
#> 
#> Initial graph
#> 
#> --- Hypothesis weights ---
#> H1: 0.5
#> H2: 0.5
#> H3: 0.0
#> H4: 0.0
#> 
#> --- Transition weights ---
#>     H1 H2 H3 H4
#>  H1  0  0  1  0
#>  H2  0  0  0  1
#>  H3  0  1  0  0
#>  H4  1  0  0  0
#> 
#> Updated graph after deleting hypotheses 2, 3
#> 
#> --- Hypothesis weights ---
#> H1: 0.5
#> H2:  NA
#> H3:  NA
#> H4: 0.5
#> 
#> --- Transition weights ---
#>     H1 H2 H3 H4
#>  H1  0 NA NA  1
#>  H2 NA NA NA NA
#>  H3 NA NA NA NA
#>  H4  1 NA NA  0

# Equivalent in ordered mode
graph_update(g, 2:3)
#> Initial and final graphs -------------------------------------------------------
#> 
#> Initial graph
#> 
#> --- Hypothesis weights ---
#> H1: 0.5
#> H2: 0.5
#> H3: 0.0
#> H4: 0.0
#> 
#> --- Transition weights ---
#>     H1 H2 H3 H4
#>  H1  0  0  1  0
#>  H2  0  0  0  1
#>  H3  0  1  0  0
#>  H4  1  0  0  0
#> 
#> Updated graph after deleting hypotheses 2, 3
#> 
#> --- Hypothesis weights ---
#> H1: 0.5
#> H2:  NA
#> H3:  NA
#> H4: 0.5
#> 
#> --- Transition weights ---
#>     H1 H2 H3 H4
#>  H1  0 NA NA  1
#>  H2 NA NA NA NA
#>  H3 NA NA NA NA
#>  H4  1 NA NA  0
#> 
#> Deletion sequence ($intermediate_graphs) ---------------------------------------
#> 
#>   Initial graph
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2: 0.5
#>   H3: 0.0
#>   H4: 0.0
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1  0  0  1  0
#>   H2  0  0  0  1
#>   H3  0  1  0  0
#>   H4  1  0  0  0
#> 
#>     Step 1: Updated graph after removing hypothesis 2
#> 
#>     --- Hypothesis weights ---
#>     H1: 0.5
#>     H2:  NA
#>     H3: 0.0
#>     H4: 0.5
#> 
#>     --- Transition weights ---
#>        H1 H2 H3 H4
#>     H1  0 NA  1  0
#>     H2 NA NA NA NA
#>     H3  0 NA  0  1
#>     H4  1 NA  0  0
#> 
#>       Step 2: Updated graph after removing hypotheses 2, 3
#> 
#>       --- Hypothesis weights ---
#>       H1: 0.5
#>       H2:  NA
#>       H3:  NA
#>       H4: 0.5
#> 
#>       --- Transition weights ---
#>          H1 H2 H3 H4
#>       H1  0 NA NA  1
#>       H2 NA NA NA NA
#>       H3 NA NA NA NA
#>       H4  1 NA NA  0
#> 
#>   Final updated graph after removing deleted hypotheses
#> 
#>   --- Hypothesis weights ---
#>   H1: 0.5
#>   H2:  NA
#>   H3:  NA
#>   H4: 0.5
#> 
#>   --- Transition weights ---
#>      H1 H2 H3 H4
#>   H1  0 NA NA  1
#>   H2 NA NA NA NA
#>   H3 NA NA NA NA
#>   H4  1 NA NA  0
#>