graphicalMCP: Graphical Multiple Comparison Procedures

Multiple comparison procedures (MCPs) control the familywise error rate in clinical trials. Graphical MCPs include many commonly used procedures as special cases; see Bretz et al. (2011) doi:10.1002/bimj.201000239 , Lu (2016) doi:10.1002/sim.6985 , and Xi et al. (2017) doi:10.1002/bimj.201600233 . This package is a low-dependency implementation of graphical MCPs which allow mixed types of tests. It also includes power simulations and visualization of graphical MCPs.

See also

Author

Maintainer: Dong Xi dong.xi1@gilead.com

Authors:

Ethan Brockmann ethan.brockmann@atorusresearch.com

Other contributors:

Gilead Sciences, Inc. [copyright holder, funder]